==> code/1. Data - comtradeapi_test.ipynb <==
##### Imports

import sys
import pandas as pd
import requests
import comtradeapicall
from datetime import date
from datetime import timedelta

### UNComtrade API Usage Examples

UN Comtrade has 3 main API tiers: 
 - Public
 - Free Subscription
 - Paid Subscription

# Set some variables
today = date.today()
yesterday = today - timedelta(days=1)
lastweek = today - timedelta(days=7)
subscription_key = open("../uncomtrade_api_key.txt").read().strip()

Available resources

comtradeapicall.listReference()

## TariffLine Data

# Call without Subscription
df = comtradeapicall.previewTarifflineData(typeCode='C', freqCode='A', clCode='HS', period='2023',
                                             reporterCode='32', cmdCode='0', flowCode='M', partnerCode=76, partner2Code=None,
                                             customsCode=None, motCode=None, maxRecords=5000, format_output='JSON',
                                             countOnly=None, includeDesc=True)

Check latest update

comtradeapicall.getLiveUpdate(subscription_key=subscription_key)

### Reporters (Countries)

reporters = pd.read_csv("../../data/UNComtrade/reporters.csv")
reporters_ids = list(reporters.id)
reporters_ids = ",".join(list(map(str,reporters_ids)))
len(reporters_ids)

Get some specific data as a test:
- Type = C (Commodities)
- Frequency = A (Annual)
- HS Nomenclature
- Year = 2022
- Reporter = 76 (Brazil)
- Commodity = 10 (Cereals)
- Flow = X (Exports)
- Partner = All partners

brazil_cereals_X = comtradeapicall.getTarifflineData(subscription_key=subscription_key, typeCode='C', freqCode='A', clCode='HS', period='2022',
                                             reporterCode='76', cmdCode='10', flowCode='X', partnerCode=reporters_ids, partner2Code=None,
                                             customsCode=None, motCode=None, maxRecords=None, format_output='JSON', 
                                             countOnly=None, includeDesc=True)

brazil_cereals_X.head()

Let's check Argentina's imports from Brazil in Cereals

argentina_cereals_M = comtradeapicall.previewTarifflineData(typeCode='C', freqCode='A', clCode='HS', period='2022',
                                             reporterCode='32', cmdCode='10', flowCode='M', partnerCode="76", partner2Code=None,
                                             customsCode=None, motCode=None, maxRecords=5000, format_output='JSON', 
                                             countOnly=None, includeDesc=True)

argentina_cereals_M.head()

See if they match

print(brazil_cereals_X[brazil_cereals_X['partnerCode'] == 32]["primaryValue"].sum())
print(argentina_cereals_M[argentina_cereals_M['partnerCode'] == 76]["primaryValue"].sum())

Let's look at the available columns

df[df.cmdCode == "02023000"].columns

Let's see what columns are non-redundant

for col in df.columns:
    if len(df[df.cmdCode == "02023000"][col].unique()) > 1:
        print(col)
