==> code/1. Data - country_id_discrepancies.ipynb <==
import pandas as pd
import os
from utils import *

baci = pd.read_csv(f"{data_paths['baci']}/BACI_HS12_Y2012_V202501.csv")
trustworthiness_scores = pd.read_csv(f"{data_paths['graphs_data']}/trustworthiness_scores.csv", dtype={"cmd": str})
atlas_df = pd.concat([pd.read_stata(f"{data_paths['atlas']}/hs12_country_country_product_year_4_2012_2016.dta"),
                pd.read_stata(f"{data_paths['atlas']}/hs12_country_country_product_year_4_2017_2021.dta"),
                pd.read_stata(f"{data_paths['atlas']}/hs12_country_country_product_year_4_2022.dta")])
atlas_countries = pd.read_csv(f"{data_paths['atlas']}/location_country.csv", encoding="latin")
reporters = pd.read_csv(f"{data_paths['uncomtrade']}/reporters.csv")

Codes in UN Comtrade not matching Atlas

trustworthiness_scores[~trustworthiness_scores.country_id.isin(atlas_countries.country_id)]["country_id"].unique()

trustworthiness_scores.loc[trustworthiness_scores.country_id == 251, ["country_id"]] = 250 # France
trustworthiness_scores.loc[trustworthiness_scores.country_id == 579, ["country_id"]] = 578 # Norway
trustworthiness_scores.loc[trustworthiness_scores.country_id == 699, ["country_id"]] = 356 # India
trustworthiness_scores.loc[trustworthiness_scores.country_id == 757, ["country_id"]] = 756 # Switzerland
trustworthiness_scores.loc[trustworthiness_scores.country_id == 842, ["country_id"]] = 840 # USA
## 490 is ASIA Others - Not present in Atlas
trustworthiness_scores[~trustworthiness_scores.country_id.isin(atlas_countries.country_id)]["country_id"].unique()

#trustworthiness_scores.to_csv("trustworthiness_scores.csv", index=False)

Codes in BACI not matching codes in ATLAS

baci[~baci.exporter.isin(atlas_countries.country_id)]["exporter"].unique()

for file in os.listdir(f"{data_paths['baci']}"):
    print(file)
    _ = pd.read_csv(f"{data_paths['baci']}/" + file, dtype={"k": str})
    if "year" in _.columns:
        continue
    _.rename(columns={"t": "year", "k": "product_id", "i": "exporter", "j": "importer", "v": "export_value", "q": "qty"}, inplace=True)
    _.loc[_.exporter == 251, ["exporter"]] = 250 # France
    _.loc[_.exporter == 579, ["exporter"]] = 578 # Norway
    _.loc[_.exporter == 699, ["exporter"]] = 356 # India
    _.loc[_.exporter == 757, ["exporter"]] = 756 # Switzerland
    _.loc[_.exporter == 842, ["exporter"]] = 840 # USA
    _.loc[_.importer == 251, ["importer"]] = 250 # France
    _.loc[_.importer == 579, ["importer"]] = 578 # Norway
    _.loc[_.importer == 699, ["importer"]] = 356 # India
    _.loc[_.importer == 757, ["importer"]] = 756 # Switzerland
    _.loc[_.importer == 842, ["importer"]] = 840 # USA
    
    print(_.columns)
    print(_[~_.exporter.isin(atlas_countries.country_id)]["exporter"].unique(), _[~_.importer.isin(atlas_countries.country_id)]["importer"].unique())
    #y = input("save?")
    #if y == "y":
    #    _.to_csv("../data/BACI_HS12_V202501/" + file)

Countries in Atlas with no bilateral recorded flows in UNCommtrade

atlas_countries.head()

d = {"Country ID": [], "Country Name": [], "Country ISO Code": [], "Reason": []}
for missing in atlas_df[~atlas_df.country_id.isin(trustworthiness_scores.country_id)]["country_id"].unique():
    missing_alpha = atlas_countries.loc[atlas_countries.country_id == missing, "iso3_code"].values[0]
    #print(missing_alpha, missing_alpha in list(reporters.reporterCodeIsoAlpha3))
    d["Country ID"].append(missing)
    d["Country Name"].append(atlas_countries.loc[atlas_countries.country_id == missing, "name_short_en"].values[0])
    d["Country ISO Code"].append(atlas_countries.loc[atlas_countries.country_id == missing, "iso3_code"].values[0])
    
    if missing_alpha in list(reporters.reporterCodeIsoAlpha3):
        d["Reason"].append("No reported data")
    else:
        d["Reason"].append("Country not present in UN Comm Trade dataset")

#pd.DataFrame(d).sort_values("Country ID").to_latex("test.txt", index=False)
