==> code/1. Data - Load BACI Data.ipynb <==
import pandas as pd
from utils import *

# BACI data

BACI provides data on bilateral trade flows for 200 countries at the product level (5000 products). Products correspond to the "Harmonized System" nomenclature (6 digit code).

Source: https://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37

### Load data

We load the data, paying special attention to the "k" column (product ID), that contains the 6 digits code. We load this as a string to preserve the leading zeros.

(From README file)
Content:

Trade flows at the year - exporter - importer - product level.

Products in Harmonized System 6-digit nomenclature.

Values in thousand USD and quantities in metric tons.

List of Variables:
- t: year
- i: exporter
- j: importer
- k: product
- v: value
- q: quantity

# More human-readable column names
colnames = {"t": "year", "i": "exporter", "j": "importer", "k": "product_id", "v": "export_value", "q": "qty"}

# Load the data for each year and rename the columns
for year in range(2012, 2024):
    df = pd.read_csv(f"{data_paths['baci']}/BACI_HS12_Y{year}_V202501.csv", dtype={"k": str})
    df.rename(columns=colnames, inplace=True) # Rename columns for easiness
    #df.to_csv(f"../data/BACI_HS12_V202501/BACI_HS12_Y{year}_V202501.csv", index=False) # Re-write
