==> code/2. Methods - First Approach - prepare-data.ipynb <==
# Prepare data for Graph NN

Imports

import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler
from utils import *

Load raw data

df = pd.concat([pd.read_stata(f"{data_paths['atlas']}/hs12_country_country_product_year_4_2012_2016.dta"),
                pd.read_stata(f"{data_paths['atlas']}/hs12_country_country_product_year_4_2017_2021.dta"),
                pd.read_stata(f"{data_paths['atlas']}/hs12_country_country_product_year_4_2022.dta")])
products = pd.read_csv(f"{data_paths['atlas']}/product_hs12.csv", dtype=str)
countries = pd.read_csv(f"{data_paths['atlas']}/location_country.csv", dtype=str)

We get a mapping from 'country_id' to 'node_id' (needed for Pytorch-Geometric)

nodes_index = countries[["country_id"]].reset_index().reset_index(names="node_id")
nodes_index = nodes_index[["node_id", "country_id"]]
nodes_index = {int(x.loc["country_id"]): int(x.loc["node_id"]) for i, x in nodes_index.iterrows()}

Let's use less granularity: Collapse 4-digit products to 2-digit

df.product_id = df.product_id.astype(str)
df = df.merge(products[["product_id", "code"]], how="left", on="product_id")
df.drop("product_id", axis=1, inplace=True)
df.rename(columns={"code": "product_id"}, inplace=True)
df.product_id = df.product_id.str[:2]
df = df.groupby(["country_id", "partner_country_id", "year", "product_id"]).agg({"export_value": "sum", "import_value": "sum", "coi": "max", "eci": "max", "pci": "max"}).reset_index()
df.head(3)

rca_drop_threshold = 0.2

# Include total volume imported for affecte importers
import_volumes = df[["country_id", "year", "product_id", "partner_country_id", "import_value"]].groupby(["country_id", "product_id", "year"]).agg({"import_value": "sum", "partner_country_id": "nunique"}).reset_index()
import_volumes.rename(columns={"country_id": "importer_id", "partner_country_id": "n_exporters"}, inplace=True)

lost_exporters = compute_lost_exporters(df=df, years=list(range(2012,2023)), import_volumes=import_volumes, drop_threshold=rca_drop_threshold)

Select Graph Layer :: Year and Product

## Select the year-product graph
year = 2022
product_id = "30"

subset = df[(df.year == year) & (df.product_id == product_id) & (df.export_value >= 0)]

## Generate the Graph Data
We need Node IDs, Node Features, Node Labels and Edge list

### Node Features & Node IDs
We start by considering 3 features: "Export Value", "COI" and "ECI"

nodes = subset[["country_id", "export_value", "coi", "eci"]].groupby("country_id").agg({"export_value": "sum", "coi": "max", "eci": "max"})
nodes.head(3)

#### RCA as a feature

##### Compute RCAs

# Compute all years country_product matrices for period
compute_country_product_matrix_dict = {}

for y in df.year.unique():
    compute_country_product_matrix_dict[y] = compute_country_product_matrix(df[df.year==y], col="export_value")

rca_dict = rca(compute_country_product_matrix_dict)

# Country RCA for a given product
rca_dict[year][[product_id]].head(3)

##### Add RCA to node feature

nodes = nodes.merge(rca_dict[year][[product_id]].rename(columns={product_id: "rca"}), left_index=True, right_index=True, how="left")
print(nodes.shape)
nodes.head(3)

### Normalize the two not normalized features

nodes[["rca"]].idxmax()

scaler = StandardScaler()
nodes[["export_value"]] = scaler.fit_transform(nodes[["export_value"]])
nodes[["rca"]] = scaler.fit_transform(nodes[["rca"]])
nodes.head(3)

## Edge List

edge_list = subset[["country_id", "partner_country_id"]]#.replace(nodes_index)
edge_list.rename(columns={"country_id": "src", "partner_country_id": "tgt"}, inplace=True)

edge_list.head()

## Nodes Labels
We want three labels:
In year *n*, for product *p*:
- 1 :: 'lost_exporter' -> Those countries that dropped ther importer status;
- 2 :: 'affected_importer' -> Those countries that are deemed affected as a result of a lost exporter;
- 0 :: 'not_affected' -> All other countries.

Compute lost_exporters dataframe

lost_exporters_subset = lost_exporters[(lost_exporters.year == year) & (lost_exporters.product_id == product_id)]
lost_exporters_subset.head(3)

# Extract affected importers
affected_importers = lost_exporters_subset[["importer_id", "affected"]].rename(columns={"importer_id": "country_id", "affected": "label"})
affected_importers.loc[:, "label"] = affected_importers.loc[:, "label"].map({True: 1, False: 0}) # Map affected importers of lost exporters with correct label
affected_importers.sort_values(["label", "country_id"], inplace=True)
affected_importers.drop_duplicates(subset="country_id", keep="last", inplace=True) # Keep nodes affected

lost_exporters_unique = lost_exporters_subset.drop_duplicates(subset="country_id", keep="first")
if len(lost_exporters_unique) > 0:
    lost_exporters_unique.loc[:, "lost_exporter"] = 1 # Append a one to identify lost exporters

affected_importers.head(3)

if len(lost_exporters_unique) > 0:
    nodes = nodes.merge(lost_exporters_unique[["country_id", "lost_exporter"]], on=["country_id"], how="left") # Finds lost exportesr
else:
    nodes["lost_exporter"] = 0
nodes = nodes.merge(affected_importers[["country_id", "label"]], on=["country_id"], how="left") # Find affected importers
nodes.fillna(0, inplace=True) # Remaining are not affected
#nodes["label"] = nodes.label_x + nodes.label_y # Unify both labels
#nodes.drop(["label_x", "label_y"], axis=1, inplace=True)
nodes.head(3)

nodes.label.unique() # Make sure there is no overlapping labels (lost_exporter AND affected importer)

#nodes[nodes.label == 3]

## Map 'country_id' to 'node_id'

#nodes.country_id = nodes.country_id.replace(nodes_index)
#nodes.rename(columns={"country_id": "node_id"}, inplace=True)
#nodes.head(3)

We get a mapping from 'country_id' to 'node_id' (needed for Pytorch-Geometric)

nodes.sort_values("country_id", inplace=True)
node_index = nodes.reset_index(names="ix")[["ix", "country_id"]]
node_index

#nodes_index = countries[["country_id"]].reset_index().reset_index(names="node_id")
#nodes_index = nodes_index[["node_id", "country_id"]]
node_index = {int(x.loc["country_id"]): int(x.loc["ix"]) for i, x in node_index.iterrows()}

edge_list.src = edge_list.src.map(node_index)
edge_list.tgt = edge_list.tgt.map(node_index)

## Save nodes and edge list

#nodes.to_csv(f"../data/nodes-{year}-{product_id}.csv", index=False)
#edge_list.to_csv(f"../data/edge_list-{year}-{product_id}.csv", index=False)

nodes["country_id"] = nodes.country_id.map(node_index)
nodes.rename(columns={"country_id": "node_id"}, inplace=True)

edge_list

nodes.groupby("label").size()
