==> code/2. Methods - trustworthiness_index-vs-discrepancy_index.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
from scipy.stats import spearmanr
from utils import *

countries = pd.read_csv(f"{data_paths['atlas']}/countries.csv", encoding="latin1")

trustworthiness = pd.read_csv(f"{data_paths['graphs_data']}/trustworthiness_scores.csv", dtype={"cmd": str})

discrepancy = []
for i, y in enumerate(range(2017, 2023), start=1):
    _ = pd.read_csv(f"{data_paths['additional']}/4.2.{i} DISCREPANCY_INDEX_H5_{y}_csv.zip", dtype={"ProductCode": str})
    _ = _[["CountryA", "ProductCode", "Year", "DI"]]
    discrepancy.append(_)
discrepancy = pd.concat(discrepancy, ignore_index=True)

trustworthiness.head()

discrepancy.head()

Convert all discrepancies to positive (ignore the 'direction' of the discrepancy)

discrepancy.loc[:, "DI"] = np.abs(discrepancy.DI)

Reduce ProductCode to single 2 digits

discrepancy.loc[:, "ProductCode"] = discrepancy["ProductCode"].str[:2]

Aggregate by Year & Commodity

discrepancy = discrepancy.groupby(["CountryA", "Year", "ProductCode"]).mean().reset_index()

Add Country ID

discrepancy = discrepancy.merge(countries[["iso_code", "country_id"]], left_on="CountryA", right_on="iso_code", how="inner")
discrepancy.head()

discrepancy.rename(columns={"Year": "year", "ProductCode": "cmd"}, inplace=True)

trust_discr = trustworthiness.merge(discrepancy[["country_id", "year", "cmd", "DI"]], on=["country_id", "year", "cmd"], how="inner")
trust_discr.head()

trust_discr.DI.describe()

spearmanr(trust_discr["DI"], trust_discr["trustworthiness"])

### Corruption Index

corruption = pd.read_csv(f"{data_paths['additional']}/4.3 Corruption Perception Index.csv")

corruption.head()

corruption = corruption.rename(columns={"ISO3": "iso_code"})

corruption = corruption.merge(countries[["iso_code", "country_id"]], on="iso_code", how="inner")

corruption = corruption.melt(id_vars=["country_id", "iso_code"], var_name="year", value_vars=[f"{x}" for x in range(2012, 2024)])
corruption["year"] = corruption["year"].astype(int)

corruption.head()

trust_discr_avg = trust_discr.groupby(["country_id", "year"])[["trustworthiness", "DI"]].mean().reset_index()
trust_discr_avg.head()

trust_discr_corrup = corruption.merge(trust_discr_avg, on=["country_id", "year"], how="inner")
trust_discr_corrup.head()

trust_discr_corrup.dropna(inplace=True)

print(spearmanr(trust_discr_corrup["trustworthiness"], trust_discr_corrup["value"]))
print(spearmanr(trust_discr_corrup["DI"], trust_discr_corrup["value"]))
