==> code/3. Results - difficult_cases.ipynb <==
Get difficult cases

from utils import *

data_paths

graph_results = {}

# For each graph typy
for graph_type in ["export", "total", "export-layered", "multi-graph-total", "multi-graph-export"]:
    print(graph_type)
    results = pd.DataFrame()

    if graph_type.startswith("multi-graph"):
        node_ids = list(pd.read_csv(f"{data_paths['graphs_data']}/{graph_type}/node_features-2022-{graph_type.removeprefix('multi-graph-')}.csv")["country_id"])
    else:
        node_ids = []

        for prod_code in [f"{x:02d}" for x in range(1, 100)]:
            if prod_code in ["77", "98", "99"]: # 77 & 98 don't exist. 99 is not present in BACI
                continue
            print(len(list(pd.read_csv(f"{data_paths['graphs_data']}/{graph_type}/node_features-2021-{prod_code}-{graph_type}.csv")["country_id"])))
            node_ids += list(pd.read_csv(f"{data_paths['graphs_data']}/{graph_type}/node_features-2021-{prod_code}-{graph_type}.csv")["country_id"])

    # For each seed
    for seed in range(1, 11):

        seed_results = pd.DataFrame()

        # For each GNN model
        for model in ["GCN", "GAT", "SAGE"]:
            _ = json.load(open(f"{data_paths['results']}/{model}/{graph_type}/{model}-{graph_type}-{seed}-predictions.json"))["predictions"]
            _ = pd.DataFrame(_)
            _.columns = [model]
            seed_results = pd.concat([seed_results, _], axis=1) # Concatenate one column per model

        # Add the seed to keep track
        seed_results["SEED"] = seed
        # Add the true label
        seed_results["LABEL"] = json.load(open(f"{data_paths['results']}/{model}/{graph_type}/{model}-{graph_type}-{seed}-predictions.json"))["labels"]
        print(seed_results)
        seed_results["node_id"] = node_ids
        results = pd.concat([results, seed_results])

    # Add the majority vote for ENSEMBLE
    results["ENSEMBLE"] = results[["GCN", "GAT", "SAGE"]].sum(axis=1)
    results["ENSEMBLE"] = results.ENSEMBLE.apply(lambda x: 1 if x > 1 else 0)
    # Check if correct
    results["ENSEMBLE"] = (results.ENSEMBLE == results.LABEL).astype(int)
    results["GCN"] = (results.GCN == results.LABEL).astype(int)
    results["GAT"] = (results.GAT == results.LABEL).astype(int)
    results["SAGE"] = (results.SAGE == results.LABEL).astype(int)
    #results.reset_index(inplace=True, drop=False, allow_duplicates=True, names="node_ix")
    results_avg = results.groupby("node_id").mean().reset_index()
    graph_results[graph_type] = results_avg

import pickle
G = pickle.load(open(f"{data_paths['graphs_data']}/export-layered/graph-2022-export-layered.pkl", "rb"))
    

len(json.load(open(f"{data_paths['results']}/GCN/export/GCN-export-{seed}-predictions.json"))["predictions"])

graph_results["total"]

# For each graph typy
for graph_type in ["export", "total", "export-layered", "multi-graph-total", "multi-graph-export"]:

    results = pd.DataFrame()

    # For each seed
    for seed in range(1, 11):

        seed_results = pd.DataFrame()

        # For each GNN model
        for model in ["GCN", "GAT", "SAGE"]:
            _ = json.load(open(f"./results/{model}/{graph_type}/{model}-{graph_type}-{seed}-predictions.json"))["predictions"]
            _ = pd.DataFrame(_)
            _.columns = [model]
            seed_results = pd.concat([seed_results, _], axis=1) # Concatenate one column per model

        # Add the seed to keep track
        seed_results["SEED"] = seed
        # Add the true label
        seed_results["LABEL"] = json.load(open(f"./results/{model}/{graph_type}/{model}-{graph_type}-{seed}-predictions.json"))["labels"]
        results = pd.concat([results, seed_results])

    # Add the majority vote for ENSEMBLE
    results["ENSEMBLE"] = results[["GCN", "GAT", "SAGE"]].sum(axis=1)
    results["ENSEMBLE"] = results.ENSAMBLE.apply(lambda x: 1 if x > 1 else 0)

    # Calculate the result metrics for each seed
    for seed in range(1,11):

        results_seed = results[results.SEED == seed]

        acc = accuracy_score(results_seed["LABEL"], results_seed["ENSEMBLE"])
        f1_macro_avg = f1_score(results_seed["LABEL"], results_seed["ENSEMBLE"], average="macro")
        f1_pos = f1_score(results_seed["LABEL"], results_seed["ENSEMBLE"], pos_label=1, average="binary")

        d = classification_report(results_seed["LABEL"], results_seed["ENSEMBLE"], output_dict=True)
        
        d["Accuracy"] = acc
        d["F1 - Avg. Macro"] = f1_macro_avg
        d["F1 - Positives"] = f1_pos
        

        with open(f"./results/ENSAMBLE/{graph_type}/ENSAMBLE-{graph_type}-{seed}-report.json", "w") as file:
            json.dump(d, file, indent=4)
